adding a linear fitter for merging

+++ main/merge.h
@@ -8,7 +8,7 @@ void logger_task(void *q);
 struct LoggerMsg
 {
     bool delayed;      // Whether the vTaskDelayUntil was delayed.
-    int32_t read_time; // usec time at end of collection
+    int64_t read_time; // usec time at end of collection
     bool imu;          // Which IMU was collected.
     int sample_count;
     lsm6dsv16x_fifo_record_t records[32]; // Up to 16 samples per read.
